Support --to-json and --from-json, merge CSV/JSON read write methods
